Added code to remove outfile if test passes. Keeps outfile if does not pass. Renamed folder containing test suite to tests instead of pytest. Other small updates.
