+++ server/config.py
@@ -2,3 +2,4 @@ import os
 
 class Configuration:
     MONGODB_URI = os.getenv('MONGODB_URI')
+
