+++ tests/README.md
@@ -1,0 +1,22 @@
+## Introduction
+
+This directory contains all tests for the backend.
+
+## Quickstart
+
+Make sure your server is up and running. Then just run 'pytest' from within the tests folder.
+
+```
+$ cd tests
+$ pytest
+```
+
+If you need some more info, add verbosity.
+```
+$ pytest -v
+```
+
+If you only want run tests within a single file, do:
+```
+$ pytest -v test_auth.py
+```
